COMP: Remove dependency on ITKV3Compatibility
Since updating the code to use BSplineTransform (instead of the deprecated
BSplineDeformableTransform) is not trivial (see 86c8966 where we reverted
our attempt), this commit introduces copy of BSplineDeformableTransform and
BSplineDeformableTransformInitializer specific to this project.

These new files are prefixed with "Niral" and uses the namespace "itk::niral::"
instead of "itk::".