COMP: Re-use ITK, SlicerExtensionModel and BRAINSTools when build extension
Now that "niral_utilities", "UnivariateEntropyMultiImageMetric" and "DTIPrep"
can all be built using a regular build of ITKv4 without requiring the option
ITKV3_COMPATIBILITY to be enabled, this commit updates the requirement
of DTIPrep Slicer extension so that the build trees provided by Slicer
can be reused.

Additionally, the build time of the Slicer extension will also be
reduced.